Improves handling of rom addresses

+++ src/newtetris/0449D0.c
@@ -1,6 +1,6 @@
 #include "common.h"
 
-extern u32 D_273A00;  // image_lut
+extern u8 D_273A00;  // image_lut
 
 ////////////////////////////////////////
 
@@ -28,8 +28,8 @@ extern u32 D_273A00;  // image_lut
 
 void func_8007FBB0(UnkStruct_21 **arg0, u32 arg1) {
   main_8004A34C_threeliner();
-  *arg0 = (UnkStruct_21 *)n64HeapAlloc(FUN_03A750_80074888_twelveliner((u32) &D_273A00, arg1));
-  FUN_03A750_800746c0_twentyliner((u32) &D_273A00, (u8 *) *arg0, arg1);
+  *arg0 = (UnkStruct_21 *)n64HeapAlloc(FUN_03A750_80074888_twelveliner(&D_273A00, arg1));
+  FUN_03A750_800746c0_twentyliner(&D_273A00, (u8 *) *arg0, arg1);
 }
 
 #pragma GLOBAL_ASM("asm/nonmatchings/newtetris/0449D0/func_8007FC14.s")
